Refactor: Move nodeUpConfigBuilder to nodemodel
Another big chunk of node code out of cloudup!
